refactored interaction code - delegated to a PlayerInteractionController component for now, to keep it out of the way during my refactor
