+++ 2D3D_UnityProject/Assets/Scenes/Ukiyo-e Environment.unity
@@ -6210,6 +6210,21 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a058c49ff778e4c4caf4ae21c99cd4ca, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &1616911466
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1616911462}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8ac78ab4d88744706b7b55f0606136a7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  player: {fileID: 1616911465}
+  orb: {fileID: 997926949}
+  mainCam: {fileID: 1524805995}
 --- !u!1 &1630674408
 GameObject:
   m_ObjectHideFlags: 0
