+++ 2D3D_UnityProject/Assets/Scripts/Player/PlayerInteractionController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ac78ab4d88744706b7b55f0606136a7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
